Complete rewrite of the standard levels. See https://github.com/mozilla/wikimo_content/issues/50 for full discussion. The new standard level page uses the same labels and expectations as the initial page, however:
- it clearly splits risk levels (the standard levels) apart from any other levels (documentation readyness, scoring,
  pass/fail tests, etc.)
- it more clearly spells out levels expectations
- it allows for indicating the level names in the label "LOW Risk", "LOW Impact", etc. vs just "LOW", which enhances
  clarity for the reader/avoids confusion
- It changes the exception time to more closely match risk instead of impact

All in all, it helps ensuring we always focus on risk (not impact, not likelihood, not scores, etc.) when using the
levels, though you can have both scores and risk (or impact and risk, or anything of your choosing as long as risk is
ALSO included).

This allows us for directly correlating measured risk between services (for ex "MEDIUM Risk" web app resulting from
observatories scans and "LOW risk" system state resulting from MIG + vuln scans vs "HIGH Impact" in an RRA mean the
complete service is currently at MEDIUM or HIGH Risk and can be represented this way in the risk heatmap).